+++ GA-AppLocker/Modules/GA-AppLocker.Scanning/Functions/Get-RemoteArtifacts.ps1
@@ -130,6 +130,8 @@ function Get-RemoteArtifacts {
                     '.vbs' { 'VBS' }
                     '.js' { 'JS' }
                     '.wsf' { 'WSF' }
+                    '.appx' { 'APPX' }
+                    '.msix' { 'APPX' }
                     default { 'Unknown' }
                 }
             }